fix set ingredients
